+++ src/relay/ir/base.cc
@@ -22,76 +22,26 @@
  * \brief The core base types for Relay.
  */
 #include <tvm/api_registry.h>
+#include <tvm/ir/type.h>
 #include <tvm/relay/base.h>
 
 namespace tvm {
 namespace relay {
 
-using tvm::IRPrinter;
 using namespace tvm::runtime;
-
-ObjectPtr<Object> GetSourceNameNode(const std::string& name) {
-  // always return pointer as the reference can change as map re-allocate.
-  // or use another level of indirection by creating a unique_ptr
-  static std::unordered_map<std::string, ObjectPtr<SourceNameNode> > source_map;
-
-  auto sn = source_map.find(name);
-  if (sn == source_map.end()) {
-    ObjectPtr<SourceNameNode> n = make_object<SourceNameNode>();
-    source_map[name] = n;
-    n->name = std::move(name);
-    return n;
-  } else {
-    return sn->second;
-  }
-}
-
-SourceName SourceName::Get(const std::string& name) {
-  return SourceName(GetSourceNameNode(name));
-}
-
-TVM_REGISTER_API("relay._make.SourceName")
-.set_body_typed(SourceName::Get);
-
-TVM_STATIC_IR_FUNCTOR(IRPrinter, vtable)
-.set_dispatch<SourceNameNode>([](const ObjectRef& ref, tvm::IRPrinter* p) {
-    auto* node = static_cast<const SourceNameNode*>(ref.get());
-    p->stream << "SourceName(" << node->name << ", " << node << ")";
-  });
-
-TVM_REGISTER_NODE_TYPE(SourceNameNode)
-.set_creator(GetSourceNameNode)
-.set_global_key([](const Object* n) {
-    return static_cast<const SourceNameNode*>(n)->name;
-  });
-
-Span SpanNode::make(SourceName source, int lineno, int col_offset) {
-  auto n = make_object<SpanNode>();
-  n->source = std::move(source);
-  n->lineno = lineno;
-  n->col_offset = col_offset;
-  return Span(n);
-}
-
-TVM_REGISTER_NODE_TYPE(SpanNode);
-
-TVM_REGISTER_API("relay._make.Span")
-.set_body_typed(SpanNode::make);
-
-TVM_STATIC_IR_FUNCTOR(IRPrinter, vtable)
-.set_dispatch<SpanNode>([](const ObjectRef& ref, tvm::IRPrinter* p) {
-    auto* node = static_cast<const SpanNode*>(ref.get());
-    p->stream << "SpanNode(" << node->source << ", " << node->lineno << ", "
-              << node->col_offset << ")";
-  });
 
 TVM_REGISTER_NODE_TYPE(IdNode);
 
 TVM_REGISTER_API("relay._base.set_span")
 .set_body_typed<void(ObjectRef, Span)>([](ObjectRef node_ref, Span sp) {
-    auto rn = node_ref.as<RelayNode>();
+  if (auto* rn = node_ref.as<RelayNode>()) {
     CHECK(rn);
     rn->span = sp;
+  } else if (auto* rn = node_ref.as<TypeNode>()) {
+    rn->span = sp;
+  } else {
+    LOG(FATAL) << "Expect Type or RelayNode ";
+  }
 });
 
 }  // namespace relay
